avoid clipping ink ripple
Add a small float value to alleviate rounding error clipping the int ripple.
This issue is not seen with a 18px ripple, but is seen with a 16px ripple.
There are other values that show the issue. I tested with sizes from 10px to 24px, which all look good now.